split out parser tests
