+++ src/pratt_parser.c
@@ -19,6 +19,7 @@
 #include <stdarg.h>
 
 #include "pratt.h"
+#include "pratt_parser.h"
 #include "pratt_debug.h"
 #include "pratt_functions.h"
 #include "pratt_scanner.h"
@@ -49,47 +50,6 @@
 
 extern AstStringArray *include_paths;
 
-void psprintf(PrattUTF8 *, const char *, ...) __attribute__((format(printf, 2, 3)));
-
-void ppAstExpression(PrattUTF8 *, AstExpression *);
-void ppAstNest(PrattUTF8 *, AstNest *);
-static void ppAstDefinitions(PrattUTF8 *, AstDefinitions *);
-static void ppAstDefinition(PrattUTF8 *, AstDefinition *);
-static void ppAstDefine(PrattUTF8 *, AstDefine *);
-static void ppAstTypeDef(PrattUTF8 *, AstTypeDef *);
-static void ppAstAlias(PrattUTF8 *, AstAlias *);
-static void ppAstUserType(PrattUTF8 *, AstUserType *);
-static void ppAstType(PrattUTF8 *, AstType *);
-static void ppAstTypeBody(PrattUTF8 *, AstTypeBody *);
-static void ppAstTypeConstructor(PrattUTF8 *, AstTypeConstructor *);
-static void ppAstTypeConstructorArgs(PrattUTF8 *, AstTypeConstructorArgs *);
-static void ppAstTypeList(PrattUTF8 *, AstTypeList *);
-static void ppAstTypeMap(PrattUTF8 *, AstTypeMap *);
-static void ppAstExpressions(PrattUTF8 *, AstExpressions *);
-static void ppAstTypeSymbols(PrattUTF8 *, AstTypeSymbols *);
-static void ppAstTypeClause(PrattUTF8 *, AstTypeClause *);
-static void ppAstTypeFunction(PrattUTF8 *, AstTypeFunction *);
-static void ppAstLookupOrSymbol(PrattUTF8 *, AstLookupOrSymbol *);
-static void ppAstLookupSymbol(PrattUTF8 *, AstLookupSymbol *);
-static void ppAstCompositeFunction(PrattUTF8 *, AstCompositeFunction *);
-static void ppAstFunction(PrattUTF8 *, AstFunction *);
-static void ppAstArgList(PrattUTF8 *, AstArgList *);
-static void ppAstTaggedArgList(PrattUTF8 *, AstTaggedArgList *);
-static void ppAstArg(PrattUTF8 *, AstArg *);
-static void ppAstNamedArg(PrattUTF8 *, AstNamedArg *);
-static void ppAstUnpack(PrattUTF8 *, AstUnpack *);
-static void ppAstUnpackStruct(PrattUTF8 *, AstUnpackStruct *);
-static void ppAstIff(PrattUTF8 *, AstIff *);
-static void ppAstLookup(PrattUTF8 *, AstLookup *);
-static void ppAstPrint(PrattUTF8 *, AstPrint *);
-static void ppAstStruct(PrattUTF8 *, AstStruct *);
-static void ppAstTaggedExpressions(PrattUTF8 *, AstTaggedExpressions *);
-
-static void ppMaybeBigInt(PrattUTF8 *, MaybeBigInt *);
-static void ppUnicodeChar(PrattUTF8 *, Character);
-static void ppHashSymbol(PrattUTF8 *, HashSymbol *);
-
-AstNest *top(PrattParser *);
 AstExpression *expr_bp(PrattParser *parser, int min_bp);
 static AstExpression *grouping(PrattRecord *, PrattParser *, AstExpression *);
 static AstExpression *prefix(PrattRecord *, PrattParser *, AstExpression *);
@@ -144,21 +104,6 @@ static AstUnpack *stringarg(PrattParser *);
 
 static PrattUnicode *PrattUTF8ToUnicode(PrattUTF8 *);
 
-void psprintf(PrattUTF8 *utf8, const char *message, ...) {
-    va_list args;
-    va_start(args, message);
-    va_list copy;
-    va_copy(copy, args);
-    size_t size = vsnprintf(NULL, 0, message, args) + 1;
-    extendPrattUTF8(utf8, utf8->size + size);
-    unsigned char *start = &utf8->entries[utf8->size];
-    vsnprintf((char *)start, size, message, copy);
-    va_end(args);
-    va_end(copy);
-    utf8->size += size;
-    utf8->size--;
-}
-
 static void addRecord(PrattTable *table, HashSymbol *tok, PrattOp prefix, PrattOp infix, PrattOp postfix, int precedence) {
     PrattRecord *record = newPrattRecord(tok, prefix, infix, postfix, precedence);
     int save = PROTECT(record);
@@ -166,7 +111,7 @@ static void addRecord(PrattTable *table, HashSymbol *tok, PrattOp prefix, PrattO
     UNPROTECT(save);
 }
 
-static PrattParser *makePrattParser() {
+PrattParser *makePrattParser() {
     PrattParser *res = newPrattParser(NULL, NULL);
     int save = PROTECT(res);
     PrattTable *table = res->rules;
@@ -439,7 +384,7 @@ static AstNamespace *parseLink(PrattParser *parser, unsigned char *file, HashSym
     return ns;
 }
 
-static PrattTrie *makePrattTrie(PrattParser *parser, PrattTrie *C) {
+PrattTrie *makePrattTrie(PrattParser *parser, PrattTrie *C) {
     HashSymbol *tok;
     Index i = 0;
     int save = PROTECT(parser); // not C because we need to have a slot for REPLACE_PROTECT
@@ -449,471 +394,6 @@ static PrattTrie *makePrattTrie(PrattParser *parser, PrattTrie *C) {
     }
     UNPROTECT(save);
     return C;
-}
-
-void ppAstNest(PrattUTF8 *dest, AstNest *nest) {
-    psprintf(dest, "{ ");
-    if (nest) {
-        if (nest->definitions) {
-            if (nest->expressions) {
-                psprintf(dest, "let ");
-                ppAstDefinitions(dest, nest->definitions);
-                psprintf(dest, "in ");
-                ppAstExpressions(dest, nest->expressions);
-            } else {
-                psprintf(dest, "namespace ");
-                ppAstDefinitions(dest, nest->definitions);
-            }
-        } else {
-            ppAstExpressions(dest, nest->expressions);
-        }
-    }
-    psprintf(dest, "}");
-}
-
-void ppAstDefinitions(PrattUTF8 *dest, AstDefinitions *definitions) {
-    while (definitions) {
-        ppAstDefinition(dest, definitions->definition);
-        psprintf(dest, "; ");
-        definitions = definitions->next;
-    }
-}
-
-static void ppAstDefinition(PrattUTF8 *dest, AstDefinition *definition) {
-    switch (definition->type) {
-        case AST_DEFINITION_TYPE_DEFINE:
-            ppAstDefine(dest, definition->val.define);
-            break;
-        case AST_DEFINITION_TYPE_TYPEDEF:
-            ppAstTypeDef(dest, definition->val.typeDef);
-            break;
-        case AST_DEFINITION_TYPE_ALIAS:
-            ppAstAlias(dest, definition->val.alias);
-            break;
-        case AST_DEFINITION_TYPE_BLANK:
-            break;
-        default:
-            cant_happen("unrecognised %s", astDefinitionTypeName(definition->type));
-    }
-}
-
-static void ppAstDefine(PrattUTF8 *dest, AstDefine *define) {
-    ppHashSymbol(dest, define->symbol);
-    psprintf(dest, " = ");
-    ppAstExpression(dest, define->expression);
-}
-
-static void ppAstTypeDef(PrattUTF8 *dest, AstTypeDef *typeDef) {
-    psprintf(dest, "typedef ");
-    ppAstUserType(dest, typeDef->userType);
-    psprintf(dest, " {");
-    ppAstTypeBody(dest, typeDef->typeBody);
-    psprintf(dest, "}");
-}
-
-static void ppAstAlias(PrattUTF8 *dest, AstAlias *alias) {
-    psprintf(dest, "alias ");
-    ppHashSymbol(dest, alias->name);
-    psprintf(dest, " = ");
-    ppAstType(dest, alias->type);
-}
-
-static void ppAstUserType(PrattUTF8 *dest, AstUserType *userType) {
-    ppHashSymbol(dest, userType->symbol);
-    if (userType->typeSymbols != NULL) {
-        psprintf(dest, "(");
-        ppAstTypeSymbols(dest, userType->typeSymbols);
-        psprintf(dest, ")");
-    }
-}
-
-static void ppAstType(PrattUTF8 *dest, AstType *type) {
-    if (type != NULL) {
-        ppAstTypeClause(dest, type->typeClause);
-        if (type->next) {
-            psprintf(dest, ", ");
-            ppAstType(dest, type->next);
-        }
-    }
-}
-
-static void ppAstTypeBody(PrattUTF8 *dest, AstTypeBody *typeBody) {
-    if (typeBody != NULL) {
-        ppAstTypeConstructor(dest, typeBody->typeConstructor);
-        if (typeBody->next) {
-            psprintf(dest, " | ");
-            ppAstTypeBody(dest, typeBody->next);
-        }
-    }
-}
-
-static void ppAstTypeConstructor(PrattUTF8 *dest, AstTypeConstructor *typeConstructor) {
-    ppHashSymbol(dest, typeConstructor->symbol);
-    ppAstTypeConstructorArgs(dest, typeConstructor->args);
-}
-
-static void ppAstTypeConstructorArgs(PrattUTF8 *dest, AstTypeConstructorArgs *typeConstructorArgs) {
-    if (typeConstructorArgs) {
-        switch (typeConstructorArgs->type) {
-            case AST_TYPECONSTRUCTORARGS_TYPE_LIST:
-                psprintf(dest, "(");
-                ppAstTypeList(dest, typeConstructorArgs->val.list);
-                psprintf(dest, ")");
-                break;
-            case AST_TYPECONSTRUCTORARGS_TYPE_MAP:
-                psprintf(dest, "{ ");
-                ppAstTypeMap(dest, typeConstructorArgs->val.map);
-                psprintf(dest, " }");
-                break;
-            default:
-                cant_happen("unrecognised %s", astTypeConstructorArgsTypeName(typeConstructorArgs->type));
-        }
-    }
-}
-
-static void ppAstTypeList(PrattUTF8 *dest, AstTypeList *typeList) {
-    if (typeList) {
-        ppAstType(dest, typeList->type);
-        if (typeList->next) {
-            psprintf(dest, ", ");
-            ppAstTypeList(dest, typeList->next);
-        }
-    }
-}
-
-static void ppAstTypeMap(PrattUTF8 *dest, AstTypeMap *typeMap) {
-    if (typeMap) {
-        ppHashSymbol(dest, typeMap->key);
-        psprintf(dest, ": ");
-        ppAstType(dest, typeMap->type);
-        if (typeMap->next) {
-            psprintf(dest, ", ");
-            ppAstTypeMap(dest, typeMap->next);
-        }
-    }
-}
-
-static void ppAstExpressions(PrattUTF8 *dest, AstExpressions *expressions) {
-    if (expressions) {
-        ppAstExpression(dest, expressions->expression);
-        psprintf(dest, "; ");
-        if (expressions->next) {
-            ppAstExpressions(dest, expressions->next);
-        }
-    }
-}
-
-static void ppAstTypeSymbols(PrattUTF8 *dest, AstTypeSymbols *typeSymbols) {
-    if (typeSymbols) {
-        ppHashSymbol(dest, typeSymbols->typeSymbol);
-        if (typeSymbols->next) {
-            psprintf(dest, ", ");
-            ppAstTypeSymbols(dest, typeSymbols->next);
-        }
-    }
-}
-
-static void ppAstTypeClause(PrattUTF8 *dest, AstTypeClause *typeClause) {
-    switch (typeClause->type) {
-        case AST_TYPECLAUSE_TYPE_INTEGER:
-            psprintf(dest, "number");
-            break;
-        case AST_TYPECLAUSE_TYPE_CHARACTER:
-            psprintf(dest, "char");
-            break;
-        case AST_TYPECLAUSE_TYPE_VAR:
-            ppHashSymbol(dest, typeClause->val.var);
-            break;
-        case AST_TYPECLAUSE_TYPE_TYPEFUNCTION:
-            ppAstTypeFunction(dest, typeClause->val.typeFunction);
-            break;
-        case AST_TYPECLAUSE_TYPE_TYPETUPLE:
-            psprintf(dest, "#(");
-            ppAstTypeList(dest, typeClause->val.typeTuple);
-            psprintf(dest, ")");
-            break;
-        default:
-            cant_happen("unrecognised %s", astTypeClauseTypeName(typeClause->type));
-    }
-}
-
-static void ppAstTypeFunction(PrattUTF8 *dest, AstTypeFunction *typeFunction) {
-    ppAstLookupOrSymbol(dest, typeFunction->symbol);
-    if (typeFunction->typeList) {
-        psprintf(dest, "(");
-        ppAstTypeList(dest, typeFunction->typeList);
-        psprintf(dest, ")");
-    }
-}
-
-static void ppAstLookupOrSymbol(PrattUTF8 *dest, AstLookupOrSymbol *lookupOrSymbol) {
-    switch (lookupOrSymbol->type) {
-        case AST_LOOKUPORSYMBOL_TYPE_LOOKUP:
-            ppAstLookupSymbol(dest, lookupOrSymbol->val.lookup);
-            break;
-        case AST_LOOKUPORSYMBOL_TYPE_SYMBOL:
-            ppHashSymbol(dest, lookupOrSymbol->val.symbol);
-            break;
-        default:
-            cant_happen("unrecognised %s", astLookupOrSymbolTypeName(lookupOrSymbol->type));
-    }
-}
-
-static void ppAstLookupSymbol(PrattUTF8 *dest, AstLookupSymbol *lookupSymbol) {
-    ppHashSymbol(dest, lookupSymbol->nsSymbol);
-    psprintf(dest, "<%d>.", lookupSymbol->nsid);
-    ppHashSymbol(dest, lookupSymbol->symbol);
-}
-
-static void ppAstLookup(PrattUTF8 *dest, AstLookup *lookup) {
-    ppHashSymbol(dest, lookup->nsSymbol);
-    psprintf(dest, "<%d>.", lookup->nsid);
-    ppAstExpression(dest, lookup->expression);
-}
-
-static void ppFunctionComponents(PrattUTF8 *dest, AstCompositeFunction *compositeFunction) {
-    if (compositeFunction) {
-        ppAstFunction(dest, compositeFunction->function);
-        psprintf(dest, " ");
-        ppFunctionComponents(dest, compositeFunction->next);
-    }
-}
-
-static void ppAstCompositeFunction(PrattUTF8 *dest, AstCompositeFunction *compositeFunction) {
-    if (compositeFunction == NULL) return;
-    if (compositeFunction->unsafe) {
-        psprintf(dest, "unsafe ");
-    }
-    psprintf(dest, "fn { ");
-    ppFunctionComponents(dest, compositeFunction);
-    psprintf(dest, "}");
-}
-
-static void ppAstFunction(PrattUTF8 *dest, AstFunction *function) {
-    psprintf(dest, "(");
-    ppAstArgList(dest, function->argList);
-    psprintf(dest, ") ");
-    ppAstNest(dest, function->nest);
-    
-}
-
-static void ppAstArgList(PrattUTF8 *dest, AstArgList *argList) {
-    if (argList) {
-        ppAstArg(dest, argList->arg);
-        if (argList->next) {
-            psprintf(dest, ", ");
-            ppAstArgList(dest, argList->next);
-        }
-    }
-}
-
-static void ppAstArg(PrattUTF8 *dest, AstArg *arg) {
-    switch (arg->type) {
-        case AST_ARG_TYPE_WILDCARD:
-            psprintf(dest, "_");
-            break;
-        case AST_ARG_TYPE_SYMBOL:
-            ppHashSymbol(dest, arg->val.symbol);
-            break;
-        case AST_ARG_TYPE_LOOKUP:
-            ppAstLookupSymbol(dest, arg->val.lookup);
-            break;
-        case AST_ARG_TYPE_NAMED:
-            ppAstNamedArg(dest, arg->val.named);
-            break;
-        case AST_ARG_TYPE_UNPACK:
-            ppAstUnpack(dest, arg->val.unpack);
-            break;
-        case AST_ARG_TYPE_UNPACKSTRUCT:
-            ppAstUnpackStruct(dest, arg->val.unpackStruct);
-            break;
-        case AST_ARG_TYPE_NUMBER:
-            ppMaybeBigInt(dest, arg->val.number);
-            break;
-        case AST_ARG_TYPE_CHARACTER:
-            ppUnicodeChar(dest, arg->val.character);
-            break;
-        case AST_ARG_TYPE_TUPLE:
-            psprintf(dest, "#(");
-            ppAstArgList(dest, arg->val.tuple);
-            psprintf(dest, ")");
-            break;
-        default:
-            break;
-    }
-}
-
-static void ppAstNamedArg(PrattUTF8 *dest, AstNamedArg *namedArg) {
-    ppHashSymbol(dest, namedArg->name);
-    psprintf(dest, " = ");
-    ppAstArg(dest, namedArg->arg);
-}
-
-static void ppAstUnpack(PrattUTF8 *dest, AstUnpack *unpack) {
-    ppAstLookupOrSymbol(dest, unpack->symbol);
-    psprintf(dest, "(");
-    ppAstArgList(dest, unpack->argList);
-    psprintf(dest, ")");
-}
-
-static void ppAstUnpackStruct(PrattUTF8 *dest, AstUnpackStruct *unpackStruct) {
-    ppAstLookupOrSymbol(dest, unpackStruct->symbol);
-    psprintf(dest, "{ ");
-    ppAstTaggedArgList(dest, unpackStruct->argList);
-    psprintf(dest, " }");
-}
-
-static void ppAstTaggedArgList(PrattUTF8 *dest, AstTaggedArgList *taggedArgList) {
-    if (taggedArgList) {
-        ppHashSymbol(dest, taggedArgList->tag);
-        psprintf(dest, ": ");
-        ppAstArg(dest, taggedArgList->arg);
-        if (taggedArgList->next) {
-            psprintf(dest, ", ");
-            ppAstTaggedArgList(dest, taggedArgList->next);
-        }
-    }
-}
-
-static void ppMaybeBigInt(PrattUTF8 *dest, MaybeBigInt *maybe) {
-    size_t size = printSizeMaybeBigInt(maybe);
-    extendPrattUTF8(dest, dest->size + size);
-    unsigned char *start = &dest->entries[dest->size];
-    size = sprintMaybeBigInt((char *)start, maybe);
-    dest->size += size;
-    dest->size--;
-}
-
-static void ppUnicodeChar(PrattUTF8 *dest, Character c) {
-    unsigned char buf[8];
-    unsigned char *end = writeChar(buf, c);
-    *end = 0;
-    psprintf(dest, "%s", buf);
-}
-
-static void ppHashSymbol(PrattUTF8 *dest, HashSymbol *symbol) {
-    psprintf(dest, "%s", symbol->name);
-}
-
-void ppAstFunCall(PrattUTF8 *dest, AstFunCall *funCall) {
-    ppAstExpression(dest, funCall->function);
-    psprintf(dest, "(");
-    for (AstExpressions *expressions = funCall->arguments; expressions != NULL; expressions = expressions->next) {
-        ppAstExpression(dest, expressions->expression);
-        if (expressions->next) {
-            psprintf(dest, ", ");
-        }
-    }
-    psprintf(dest, ")");
-}
-
-void ppAstCharacter(PrattUTF8 *dest, Character c) {
-    unsigned char buffer[8];
-    unsigned char *end = writeChar(buffer, c);
-    *end = 0;
-    psprintf(dest, "'%s'", buffer);
-}
-
-void ppAstTuple(PrattUTF8 *dest, AstExpressions *expressions) {
-    psprintf(dest, "#(");
-    while (expressions) {
-        ppAstExpression(dest, expressions->expression);
-        if (expressions->next)
-            psprintf(dest, ", ");
-        expressions = expressions->next;
-    }
-    psprintf(dest, ")");
-}
-
-static void ppAstIff(PrattUTF8 *dest, AstIff *iff) {
-    psprintf(dest, "if (");
-    ppAstExpression(dest, iff->test);
-    psprintf(dest, ") ");
-    ppAstNest(dest, iff->consequent);
-    psprintf(dest, " else ");
-    ppAstNest(dest, iff->alternative);
-}
-
-static void ppAstPrint(PrattUTF8 *dest, AstPrint *print) {
-    psprintf(dest, "print(");
-    ppAstExpression(dest, print->exp);
-    psprintf(dest, ")");
-}
-
-static void ppAstStruct(PrattUTF8 *dest, AstStruct *structure) {
-    ppAstLookupOrSymbol(dest, structure->symbol);
-    psprintf(dest, "{ ");
-    ppAstTaggedExpressions(dest, structure->expressions);
-    psprintf(dest, " }");
-}
-
-static void ppAstTaggedExpressions(PrattUTF8 *dest, AstTaggedExpressions *taggedExpressions) {
-    if (taggedExpressions) {
-        ppHashSymbol(dest, taggedExpressions->tag);
-        psprintf(dest, ": (");
-        ppAstExpression(dest, taggedExpressions->expression);
-        psprintf(dest, ")");
-        if (taggedExpressions->next) {
-            psprintf(dest, ", ");
-            ppAstTaggedExpressions(dest, taggedExpressions->next);
-        }
-    }
-}
-
-void ppAstExpression(PrattUTF8 *dest, AstExpression *expr) {
-    switch (expr->type) {
-        case AST_EXPRESSION_TYPE_NUMBER:
-            ppMaybeBigInt(dest, expr->val.number);
-            break;
-        case AST_EXPRESSION_TYPE_SYMBOL:
-            psprintf(dest, "%s", expr->val.symbol->name);
-            break;
-        case AST_EXPRESSION_TYPE_FUNCALL:
-            ppAstFunCall(dest, expr->val.funCall);
-            break;
-        case AST_EXPRESSION_TYPE_CHARACTER:
-            ppAstCharacter(dest, expr->val.character);
-            break;
-        case AST_EXPRESSION_TYPE_TUPLE:
-            ppAstTuple(dest, expr->val.tuple);
-            break;
-        case AST_EXPRESSION_TYPE_FUN:
-            ppAstCompositeFunction(dest, expr->val.fun);
-            break;
-        case AST_EXPRESSION_TYPE_IFF:
-            ppAstIff(dest, expr->val.iff);
-            break;
-        case AST_EXPRESSION_TYPE_NEST:
-            ppAstNest(dest, expr->val.nest);
-            break;
-        case AST_EXPRESSION_TYPE_BACK:
-            psprintf(dest, "back");
-            break;
-        case AST_EXPRESSION_TYPE_ENV:
-            psprintf(dest, "env");
-            break;
-        case AST_EXPRESSION_TYPE_LOOKUP:
-            ppAstLookup(dest, expr->val.lookup);
-            break;
-        case AST_EXPRESSION_TYPE_PRINT:
-            ppAstPrint(dest, expr->val.print);
-            break;
-        case AST_EXPRESSION_TYPE_STRUCTURE:
-            ppAstStruct(dest, expr->val.structure);
-            break;
-        case AST_EXPRESSION_TYPE_ASSERTION:
-            psprintf(dest, "assert(");
-            ppAstExpression(dest, expr->val.assertion);
-            psprintf(dest, ")");
-            break;
-        case AST_EXPRESSION_TYPE_ERROR:
-            psprintf(dest, "error(");
-            ppAstExpression(dest, expr->val.error);
-            psprintf(dest, ")");
-            break;
-        default:
-            cant_happen("unexpected %s", astExpressionTypeName(expr->type));
-    }
 }
 
 static AstExpression *makePrattBinary(ParserInfo I, HashSymbol *op, AstExpression *lhs, AstExpression *rhs) {
@@ -1908,71 +1388,4 @@ AstExpression *expr_bp(PrattParser *parser, int min_bp) {
     LEAVE(expr_bp);
     UNPROTECT(save);
     return lhs;
-}
-
-static void test(PrattParser *parser, PrattTrie *trie, char *expr) {
-    printf("%-30s ", expr);
-    NEWLINE();
-    parser->lexer = makePrattLexer(trie, "test", expr);
-    AstNest *result = top(parser);
-    int save = PROTECT(result);
-    if (parser->lexer->bufList != NULL) {
-        PrattToken *tok = next(parser->lexer);
-        errorAt(tok, "unconsumed tokens");
-    }
-    PrattUTF8 *dest = newPrattUTF8();
-    PROTECT(dest);
-    ppAstNest(dest, result);
-    printf("%s", dest->entries);
-    printf("\n");
-    UNPROTECT(save);
-}
-
-int main(int argc __attribute__((unused)), char *argv[] __attribute__((unused))) {
-    initProtection();
-    PrattParser *p = makePrattParser();
-    PROTECT(p);
-    PrattTrie *t = makePrattTrie(p, NULL);
-    PROTECT(t);
-    test(p, t, "1");
-    test(p, t, "5!");
-    test(p, t, "1i");
-    test(p, t, "Σ");
-    test(p, t, "'Σ'");
-    test(p, t, "a123");
-    test(p, t, "1 + 2");
-    test(p, t, "1 <=> 2 <=> 3");
-    test(p, t, "1 + 2 * 3");
-    test(p, t, "1 * 2 + 3");
-    test(p, t, "an and android");
-    test(p, t, "1 * 2 * 3");
-    test(p, t, "1 . 2 . 3");
-    test(p, t, "- 1 . - 2 . 3");
-    test(p, t, "--1 * 2");
-    test(p, t, "--1 * 2!");
-    test(p, t, "1 * ((2 + 3))");
-    test(p, t, "a -> b -> c");
-    test(p, t, "(a -> b) -> c");
-    test(p, t, "<a @ b");
-    test(p, t, "a @ b < c");
-    test(p, t, "1 then 2 then 3");
-    test(p, t, "a(b)");
-    test(p, t, "a(b, c)");
-    test(p, t, "#(b)");
-    test(p, t, "#(b, c)");
-    test(p, t, "#a + b");
-    test(p, t, "a + #b");
-    test(p, t, "a #b");
-    test(p, t, "a @ b @@ c @ d");
-    test(p, t, "123456789012345678901234567890");
-    test(p, t, "12345678901234567890123456789i");
-    test(p, t, "12345.6789i");
-    test(p, t, "let fn i(x) { x } in i(0)");
-    test(p, t, "if (a > 2) { 3 } else { 4 }");
-    test(p, t, "let x = 1 then 2 then 3; in x");
-    test(p, t, "let typedef named_list(#t) { nl(str, list(#t)) } in nl ");
-
-    // test("1 * ((2 + 3[4 + 5]))");
-    // test("aa = bb = 3 ? 4 ? 5 : 6 : 7 ? 8 : 9");
-    return 0;
 }
